Threatviz_Agentcore/pyproject.toml
[project]
name = "smartvm"
version = "0.1.0"
description = "this tool can be used to Imagine a threatmodel out of a CVE"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "agentcore>=0.1.6",
    "aws>=0.2.5",
    "beautifulsoup4>=4.14.3",
    "bedrock-agentcore-starter-toolkit>=0.2.5",
    "chroma>=0.2.0",
    "colorama>=0.4.6",
    "faiss-cpu>=1.13.0",
    "langchain-anthropic>=1.3.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "langchain[aws]>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-aws>=1.0.1",
    "pypdf>=6.4.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.2",
]
